Make stringbuffer length an int
Avoid warnings from MSVC by using an int for size instead of
size_t. While size_t is probably the more correct type to use we
should get rid of the stringbuffer class anyway so just satisfy MSVC
for now.

+++ src/KaperCanvas.cpp
@@ -621,7 +621,7 @@ void KaperCanvas::DrawString(int iNumbers, int x, int y, bool bBold, int iColor)
     for (int i = 0; i < 4; i++) {
       iYoffset = 0;
 
-      if (static_cast<int>(m_stringBuffer->length()) <= i) {
+      if (m_stringBuffer->length() <= i) {
         break;
       }
